Read PLC weight as 32-bit IEEE 754 float (2 words)
- MXPlc: add ReadFloat() - reads consecutive Dn + Dn+1 registers,
  combines as (high<<16)|low and converts to IEEE 754 float
- Form1.WeightTimer_Elapsed: ReadShort -> ReadFloat, 0-check uses 0.0f

+++ PoongSan_Angang_BCR/Form1.cs
@@ -1236,9 +1236,9 @@ namespace PoongSan_Angang_BCR
             {
                 if (string.IsNullOrEmpty(address)) continue; // 미설정 탄 → 건너뜀
 
-                short rawValue;
-                if (m_VasimPlatform.m_mxPlc.ReadShort(address, out rawValue) != 0) continue; // 읽기 실패 → 무시
-                if (rawValue == 0) continue; // 값이 0 → 무시 (요구사항)
+                float rawValue;
+                if (m_VasimPlatform.m_mxPlc.ReadFloat(address, out rawValue) != 0) continue; // 읽기 실패 → 무시
+                if (rawValue == 0.0f) continue; // 값이 0 → 무시 (요구사항)
 
                 double weight = rawValue; // 단위: grain
                 if (weight < _currentWeightMin || weight > _currentWeightMax)
